Improve screen cloning when there are more than two monitors, or only one.
